[Fix] bot action for more strict data struct (tools)

+++ tools/bot-actions/BotTargetListGUI.cpp
@@ -550,7 +550,7 @@ std::vector<std::string> BotTargetList::contentToGUI_internal(const CatchChallen
             for(auto it = blockObject->botsFight.begin();it!=blockObject->botsFight.cend();++it)
             {
                 const std::pair<uint8_t,uint8_t> &point=it->first;
-                const std::vector<uint32_t> &botsFightList=it->second;
+                const std::vector<uint16_t> &botsFightList=it->second;
                 unsigned int index=0;
                 while(index<botsFightList.size())
                 {
@@ -741,7 +741,7 @@ std::vector<std::string> BotTargetList::contentToGUI_internal(const CatchChallen
         {
             for(auto it = blockObject->shops.begin();it!=blockObject->shops.cend();++it)
             {
-                const std::vector<uint32_t> &shops=it->second;
+                const std::vector<uint16_t> &shops=it->second;
                 unsigned int index=0;
                 while(index<shops.size())
                 {
